Refactor: コンポーネントをatomicデザインのベースに分類
