+++ app/javascript/packs/application.js
@@ -12,11 +12,11 @@ ReactOnRails.register({ App });
 // const images = require.context("../images", true);
 // const imagePath = (name) => images(name, true);
 
-const setHeight = () => {
-  let vh = window.innerHeight * 0.01;
-  document.documentElement.style.setProperty("--vh", `${vh}px`);
-}
-
 (() => {
+  const setHeight = () => {
+    let vh = window.innerHeight * 0.01;
+    document.documentElement.style.setProperty("--vh", `${vh}px`);
+  }
+
   setHeight();
 })();
