+++ app/javascript/packs/components/pages/SelectDrink.tsx
